[refactoring/daengle-27] 각 모듈 알림 연결 추가
