+++ daengle-payment-api/build.gradle
@@ -17,6 +17,7 @@ dependencies {
     implementation(project(':daengle-domain'))
     implementation(project(':daengle-auth'))
     implementation(project(':daengle-persistence'))
+    implementation(project(':daengle-notification'))
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
     compileOnly 'org.projectlombok:lombok'
