+++ daengle-vet-api/src/main/java/ddog/vet/presentation/notification/NotificationController.java
@@ -1,0 +1,22 @@
+package ddog.vet.presentation.notification;
+
+import ddog.notification.application.NotificationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/notify")
+public class NotificationController {
+
+    private final NotificationService notificationService;
+
+    @GetMapping("/{userId}")
+    public SseEmitter connectSseEmitter(@PathVariable("userId") Long userId) {
+        return notificationService.connect(userId);
+    }
+}
